packages/model_gateway/pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "scriptkiddie-model-gateway"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["requests>=2.32.3", "pydantic>=2.7.4", "pyyaml>=6.0.1"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/model_gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
